+++ etc/deploy.sh
@@ -9,7 +9,7 @@
 # This value indicates the version number when this file was last changed: 
 # it does not necessarily reflect deploy's current version number.
 # DO NOT EDIT THIS NUMBER OR YOU MAY BLOW SOMETHING UP
-PROJECT_VERSION="3.7"
+PROJECT_VERSION="3.7.1"
 
 
 ###############################################################################