+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - If a project is using `mina` and configured to build assets on every deployment, `deploy` will skip the redunant build step
 - Test emails now contain more project information
-- Improved readability of HTML log files
+- Improved readability of log files
 - Improved unit testing
 - Cleaned up language for more consistency
 ### Fixed